+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/investigation/rest/InvestigationsController.java
@@ -47,7 +47,13 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 import qualitynotification.base.response.QualityNotificationIdResponse;
 import qualitynotification.investigation.response.InvestigationResponse;
 
@@ -115,7 +121,13 @@ public class InvestigationsController {
     public QualityNotificationIdResponse investigateAssets(@RequestBody @Valid StartQualityNotificationRequest request) {
         log.info(API_LOG_START + " with params: {}", request);
         return new QualityNotificationIdResponse(investigationService.start(
-                request.getPartIds(), request.getDescription(), request.getTargetDate(), request.getSeverity().toDomain(), request.isAsBuilt()).value());
+                        request.getPartIds(),
+                        request.getDescription(),
+                        request.getTargetDate(),
+                        request.getSeverity().toDomain(),
+                        request.getReceiverBpn(),
+                        request.isAsBuilt())
+                .value());
     }
 
     @Operation(operationId = "getCreatedInvestigations",