+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/application/request/StartQualityNotificationRequest.java
@@ -52,4 +52,6 @@ public class StartQualityNotificationRequest {
     private QualityNotificationSeverityRequest severity;
     @ApiModelProperty(example = "true")
     private boolean isAsBuilt = true;
+    @ApiModelProperty(example = "BPN00001123123AS")
+    private String receiverBpn;
 }